+++ protocols/dcutr/src/behaviour.rs
@@ -57,9 +57,11 @@ pub enum Event {
     },
     DirectConnectionUpgradeSucceeded {
         remote_peer_id: PeerId,
+        connection_id: ConnectionId,
     },
     DirectConnectionUpgradeFailed {
         remote_peer_id: PeerId,
+        connection_id: ConnectionId,
         error: Error,
     },
 }
@@ -145,6 +147,7 @@ impl Behaviour {
             self.queued_events.extend([ToSwarm::GenerateEvent(
                 Event::DirectConnectionUpgradeFailed {
                     remote_peer_id: peer_id,
+                    connection_id: failed_direct_connection,
                     error: Error::Dial,
                 },
             )]);
@@ -257,6 +260,7 @@ impl NetworkBehaviour for Behaviour {
             self.queued_events.extend([ToSwarm::GenerateEvent(
                 Event::DirectConnectionUpgradeSucceeded {
                     remote_peer_id: peer,
+                    connection_id,
                 },
             )]);
         }
@@ -294,6 +298,7 @@ impl NetworkBehaviour for Behaviour {
                 self.queued_events.push_back(ToSwarm::GenerateEvent(
                     Event::DirectConnectionUpgradeFailed {
                         remote_peer_id: event_source,
+                        connection_id,
                         error: Error::Handler(error),
                     },
                 ));
@@ -314,6 +319,7 @@ impl NetworkBehaviour for Behaviour {
                 self.queued_events.push_back(ToSwarm::GenerateEvent(
                     Event::DirectConnectionUpgradeFailed {
                         remote_peer_id: event_source,
+                        connection_id: relayed_connection_id,
                         error: Error::Handler(error),
                     },
                 ));